move generateKubeconfigs to kubeconfig.go
