+++ pkg/cluster/kubeconfig.go
@@ -4,6 +4,7 @@ package cluster
 // Licensed under the Apache License 2.0.
 
 import (
+	"context"
 	"crypto/x509"
 	"crypto/x509/pkix"
 	"strings"
@@ -15,7 +16,9 @@ import (
 	"github.com/openshift/installer/pkg/asset/tls"
 	clientcmdv1 "k8s.io/client-go/tools/clientcmd/api/v1"
 
+	"github.com/Azure/ARO-RP/pkg/api"
 	"github.com/Azure/ARO-RP/pkg/cluster/graph"
+	"github.com/Azure/ARO-RP/pkg/util/stringutils"
 )
 
 // generateAROServiceKubeconfig generates additional admin credentials and a
@@ -35,6 +38,54 @@ func (m *manager) generateAROSREKubeconfig(pg graph.PersistedGraph) (*kubeconfig
 // kubeconfig for ARO User, based on the admin kubeconfig found in the graph.
 func (m *manager) generateUserAdminKubeconfig(pg graph.PersistedGraph) (*kubeconfig.AdminInternalClient, error) {
 	return generateKubeconfig(pg, "system:admin", nil, tls.ValidityOneYear)
+}
+
+func (m *manager) generateKubeconfigs(ctx context.Context) error {
+	resourceGroup := stringutils.LastTokenByte(m.doc.OpenShiftCluster.Properties.ClusterProfile.ResourceGroupID, '/')
+	account := "cluster" + m.doc.OpenShiftCluster.Properties.StorageSuffix
+
+	pg, err := m.graph.LoadPersisted(ctx, resourceGroup, account)
+	if err != nil {
+		return err
+	}
+
+	var adminInternalClient *kubeconfig.AdminInternalClient
+	err = pg.Get(&adminInternalClient)
+	if err != nil {
+		return err
+	}
+
+	aroServiceInternalClient, err := m.generateAROServiceKubeconfig(pg)
+	if err != nil {
+		return err
+	}
+	aroSREInternalClient, err := m.generateAROSREKubeconfig(pg)
+	if err != nil {
+		return err
+	}
+	aroUserInternalClient, err := m.generateUserAdminKubeconfig(pg)
+	if err != nil {
+		return err
+	}
+
+	m.doc, err = m.db.PatchWithLease(ctx, m.doc.Key, func(doc *api.OpenShiftClusterDocument) error {
+		// used for the SAS token with which the bootstrap node retrieves its
+		// ignition payload
+		var t time.Time
+		if doc.OpenShiftCluster.Properties.Install.Now == t {
+			// Only set this if it hasn't been set already, since it is used to
+			// create values for signedStart and signedExpiry in
+			// deployResourceTemplate, and if these are not stable a
+			// redeployment will fail.
+			doc.OpenShiftCluster.Properties.Install.Now = time.Now().UTC()
+		}
+		doc.OpenShiftCluster.Properties.AdminKubeconfig = adminInternalClient.File.Data
+		doc.OpenShiftCluster.Properties.AROServiceKubeconfig = aroServiceInternalClient.File.Data
+		doc.OpenShiftCluster.Properties.AROSREKubeconfig = aroSREInternalClient.File.Data
+		doc.OpenShiftCluster.Properties.UserAdminKubeconfig = aroUserInternalClient.File.Data
+		return nil
+	})
+	return err
 }
 
 func generateKubeconfig(pg graph.PersistedGraph, commonName string, organization []string, validity time.Duration) (*kubeconfig.AdminInternalClient, error) {
